Project without AppShell and with TabbedPages including Icon (#6)
* Create subfolder for the project and move files.
* New .net8 MAUI project to test removing appshell
* Get rid of appshell and open ContentPage
* Add content page two
* Set TabPageExample as main page and add related files

Modified App.xaml.cs to set TabPageExample as the main page.
Updated NonAppShellPractice.csproj to include TabPageExample.xaml.
Added TabPageExample.xaml defining a TabbedPage with two content pages.
Added TabPageExample.xaml.cs defining the TabPageExample class.

* Add IconImageSource to ContentPage Title attribute

The `Title` attribute of the `ContentPage` element was modified to include an `IconImageSource` attribute. The `Title` attribute value remains the same ("This is page one!"), but the `IconImageSource` attribute with the value "phone.png" was added.